fix all uses of float for money related data. A float looses precision

+++ day1/src/Hotel.Payments/Domain/Payment.cs
@@ -3,14 +3,14 @@ using static Hotel.Payments.Domain.PaymentEvents;
 
 namespace Hotel.Payments.Domain {
     public class Payment : Aggregate<PaymentState, PaymentId> {
-        public void ProcessPayment(PaymentId paymentId, string bookingId, float amount, string method, string provider) {
+        public void ProcessPayment(PaymentId paymentId, string bookingId, decimal amount, string method, string provider) {
             Apply(new PaymentRecorded(paymentId, bookingId, amount, method, provider));
         }
     }
 
     public record PaymentState : AggregateState<PaymentState, PaymentId> {
         public string BookingId { get; init; }
-        public float  Amount    { get; init; }
+        public decimal Amount    { get; init; }
     }
 
     public record PaymentId(string Value) : AggregateId(Value);
